Modified CPLOPConnection to have an additional query: All pyroprints in CPLOP.
Additionally, modified InputDialog to have an action for selecting a
pyroprint data set, and for populating the data set text field.

Finally, I separated the actionlistener for the button for choosing data
from CPLOP so that it is easier to read, and appropriately separated into
various methods.

ALSO IMPORTANT TO NOTE: Developing on a windows machine, the way to get
the MYSQL JDBC driver loaded, you have to use cmd (not BASH!) and the
classpath should be ".;<path_to_jdbc_driver>" matter of fact I will set an
environment variable for it now and maybe that will fix it