Made all of the base metrics classes (including ___AverageMetric) compliant with the new approach. This means that PyroprintMetric is given 2 pyroprints to compare, ITSRegionMetric is given 2 ITSRegions to compare, etc. Additionally, Debug and Error outputs are appropriately replaced with calls to Logger.debug and Logger.error, respectively.
All reset() calls are made in result() and in
some cases there may be additional consolidations of some of these methods in
the future, although I think I actually prefer the apply() -> result() process.
It is prone to error but may yield some extra flexibility in the future.

Isolate now contains a HashMap of regions mapping region name (string) ->
region object (ITSRegion). This is so that when the IsolateMetric is given two
isolates to compare, it goes through each of it's ITSRegionMetrics and
the region that metric is meant to be applied to is then pulled from the
Isolate regionMap by name. Otherwise it would be inefficient and unreliable to
compare ITSRegions for isolates.

Additionally, I have added errorcodes to each metric including DataMetric (the
interface). This is so that I no longer have to ruin my code appearance by
being explicit with unwrapping and wrapping double values in Double objects. I
can just pass doubles around and check error codes (getError() or mErrorCode)
in order to see if the received value is a valid one. I will have to be sure
that the GUI somehow flags values as incorrect should any error codes be set so
that it is apparent when computed values should be disregarded.