Propagated changes up through AgglomerativeClusterer. Renamed PartitionedAgglomerative to be OhClustering but have not modified the file otherwise.
I changed HierarchicalCluster to be HCluster (HierarchicalCluster was too long
of a name) and I have then proceeded to make HierarchicalClusterer,
AgglomerativeClusterer, and eventually OHClustering to use HClusters instead of
Clusters.

Also removing the use of ClusterAnalyzer in each of the files. Going to see if
I can't simplify that process.