Changed Makefile to use -Xlint when compiling.
Found the bug preventing clusters from joining which was the hardcoded upper
threshold was in percent, not in decimal value.

Clusters are now joining, and distances seem to be correctly calculated between
individual isolates.

Distances between clusters with multiple isolates needs to be double checked
and the hardcoded upper threshold for isolate similarity needs to be addressed.