Renamed OHClustering to OHClusterer for consistency.
modified the querying code in CPLOPConnection for getData so that pyroprints
are primarily ordered by isoID (actually separates pyroprints from different
isolates), then pyroID (ensures that different pyroprints for the same isolate
are appropriately separated), then position (so that the pyroprint peaks is in
order).

modified construct isolates in InputDialog to more correctly construct the
pyroprints for an Isolate and to add the pyroprints to the correct isolate
(checking which region of the isolate is the same but previously I was not
maintaining multiple pyroprints for a particular isolate).

Did some insignificant changes on DataQueryButtonListern and ClusterFileDialog
primarily because I'm now running java 7 on both desktop and laptop systems and
so JComboBox has a generic (in 7) and so this was to appease Xlint.

modified toString method in Pyroprint for easier debugging (now prints out the
name of the pyroprint including the pyroID and well)