Added ClusterInterface which is going to be a generic interface to the clustering program. I should make this more API based perhaps but this should be fine for now. It's a quick and easy single-point entry into calling clustering.
ClusterResults is a class for containing clusters. Not sure what to put in it
for now but it's not a big deal, we can test it and add to it as necessary.

modified CPLOPConnection to include a debug outprint of what it's query is...
but that should be removed.. >__>

modified makefile to include targets for the generic interface