I got OHClust! to work correctly!
I have since made changes in order to set the test harness up to iterate on a
set of configuration files. The configuration files will specify an ontology to
use (if necessary) and other parameters for each evaluation configuration. I
still am considering how data will be queried and stored to make it easy to
analyze correctness/accuracy/performance of all clustering algorithms and also
how updates will be handled (ensure that only new data in an update, and how to
randomize where in the ontology data in an update will go).