When adding labels to records (in SPAMEvaluation.java, ClusterInterface.java, and InputDialog.java) the label is trimmed because some values in the database have trailing whitespace (specifically 'Human ' in commonName). This was making ontological clustering miss out on a lot of data at the time.
OHClusterer has been cleaned of various comments and debug statements and has
also been updated so that if there is no partition that a data point should be
added then it is added at it's current position in the ontology. That is, for
data points that have no place anywhere in the ontology, they are placed in the
root node and will only join cluster results once every sub-ontology has been
clustered. Other data points will percolate down the ontological structure as
far as they can go until they no longer match any partitions or they are at a
leaf partition.

some changes to the configurations have been made: it seems that the 16S-23S
ITS region should use 95 dispensations, also there were some filenames changed.

adding the strain ddl for tracking.

Also removed nohup.out because it totally does not belong in the repository.

removed some comments and debugging from Ontology and OntologyTerm

Also, had to make some updates to CPLOPConnection. getData no longer does a
pyroID filter ('where pyroID in (<pyroID List>)') because it seemed
superfluous. It may not actually be, but at the very least having that in
getData and not in getRandomData made it nearly impossible to make
constructOntologicalQuery successfully build a query for ontological data for
both.