Did some syncing between the eclipse branch and the evaluation branch. This is to get rid of various unnecessary things (such as unused imports, etc.) and to update this branch per the changes I made in eclipse while using the debugger.
Also, important note: pyroPrintedDate in kevin's crap tables are varchars not
date fields, aka when I use them to try and filter the pyroPrintedDate field
(an actual date field) in the test_pyroprints table then it doesn't match
anything because it's trying to match strings and dates. So I had to blacklist
that attribute in order to correctly get data out of the database when using
pyroPrintedDate (which pretty much means I should not be using date at all)

some updates to OHClusterer. There are still some cases I need to think through
and it sucks that I am working these things out so last minute. sucks to not
coming across these obscure cases earlier in working with the clustering.