removing .class files
There are still references to Cluster.java. I'm not sure if I removed this
class previously for a reason, or if it was accidentally deleted. Adding a
reasonable version for now to be used until I remember why it was deleted in
the first place (maybe adding it will allow me to view log history on the file)